# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "email": "contact@touri-touri.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/medias": {
            "get": {
                "tags": [
                    "Medias"
                ],
                "summary": "Get list of medias",
                "description": "Returns list of medias",
                "operationId": "getMediasList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Media"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Medias"
                ],
                "summary": "Store new media",
                "description": "Returns media data",
                "operationId": "storeMedia",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "path"
                                ],
                                "properties": {
                                    "path": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Media created successfully"
                                        },
                                        "media": {
                                            "$ref": "#/components/schemas/Media"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/medias/{id}": {
            "get": {
                "tags": [
                    "Medias"
                ],
                "summary": "Get media information",
                "description": "Returns media data",
                "operationId": "getMediaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of media to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Media found successfully"
                                        },
                                        "media": {
                                            "$ref": "#/components/schemas/Media"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Medias"
                ],
                "summary": "Delete existing media",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteMedia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of media to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Media deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Media": {
                "title": "Media",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Media 1"
                    },
                    "path": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "http://localhost:8000/storage/medias/1.jpg"
                    },
                    "type": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "image"
                    },
                    "is_main": {
                        "type": "boolean",
                        "readOnly": "true",
                        "example": "true"
                    },
                    "created_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    }
}